Alternative to initial proposal
This commit is a work in progress as I've been unable to test those
changes against keycloak.

Rather than configuring the oauth2 client, Spring Boot's OAuth2 facility
is used with a dedicated "keycloak" provider. To avoid copy/pasting the
base url, a temporary property is used in configuration.

The use of lombok has been removed as it makes hard for anyone willing
to try out the code in their IDE if they haven't installed the plugin.

A default user is still created on startup but that should not be the
case in the next milestone. Besides configuration, the sample is now
much simpler IMO.